registry/requirements.txt
Django==3.0.8
django-filter==2.3.0
django-autocomplete-light==3.5.1